created: the playground component
