+++ src/Main/Main.css
@@ -2,29 +2,7 @@
   display: flex;
   height: 100%;
 }
-
-.main__playground {
-  width: 100%;
-  height: 100%;
-  border: 1px solid black;
-  border-top: 0;
-  position: relative;
-  overflow: hidden;
-}
-
 .main__results {
   width: auto;
   height: 100%;
-}
-
-.cube {
-  position: absolute;
-  width: 40px;
-  height: 40px;
-  background-color: pink;
-  cursor: pointer;
-}
-
-.cube:hover {
-  border: 1px solid red;
 }
