+++ src/Main/Playground/Playground.css
@@ -1,0 +1,20 @@
+.playground {
+  width: 100%;
+  height: 100%;
+  border: 1px solid black;
+  border-top: 0;
+  position: relative;
+  overflow: hidden;
+}
+
+.cube {
+  position: absolute;
+  width: 40px;
+  height: 40px;
+  background-color: pink;
+  cursor: pointer;
+}
+
+.cube:hover {
+  border: 1px solid red;
+}
